Moved task to public package
Extracted out the implementation to a separate class
