+++ src/main/groovy/org/shipkit/gradle/init/InitTravisTask.java
@@ -1,10 +1,11 @@
-package org.shipkit.gradle;
+package org.shipkit.gradle.init;
 
 import org.gradle.api.DefaultTask;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
 import org.gradle.api.tasks.OutputFile;
 import org.gradle.api.tasks.TaskAction;
+import org.shipkit.internal.gradle.init.tasks.InitTravis;
 import org.shipkit.internal.notes.util.IOUtil;
 
 import java.io.File;
@@ -15,19 +16,10 @@ import java.io.InputStream;
  */
 public class InitTravisTask extends DefaultTask {
 
-    private final static Logger LOG = Logging.getLogger(InitTravisTask.class);
-
     @OutputFile private File outputFile;
 
     @TaskAction public void initTravis() {
-        if (outputFile.exists()) {
-            LOG.lifecycle("  {} - file exists, skipping generation of '{}'.", this.getPath(), outputFile.getName());
-            return;
-        }
-        InputStream resource = this.getClass().getClassLoader().getResourceAsStream("template.travis.yml");
-        String template = IOUtil.readFully(resource);
-        IOUtil.writeFile(outputFile, template);
-        LOG.lifecycle("  {} - generated default '{}', don't forget to check it in to your source control!", this.getPath(), outputFile.getName());
+        new InitTravis().initTravis(this);
     }
 
     /**
